+++ app/src/main/java/com/example/braulyse/model/DensityMeasurement.kt
@@ -1,6 +1,6 @@
 package com.example.braulyse.model
 
 data class DensityMeasurement(
-    val initialDensity: Float,
-    val finalDensity: Float
+    val initialDensity: Double,
+    val finalDensity: Double
 )
